QDoc: Qualify non-dependent types of a class template with its declaration
QDoc uses Clang's APIs to parse the source code of a project both to
extract the user-provided documentation and to perform certain sanity
chekcs on that documentation.

Furthermore it uses those same APIs to retrieve a user stringified
representation of certain documented elements from the project's code
that are shown to the end user.

For example, it might retrieve from the underlying code the name of the
types of the arguments of a function, which are then shown in the output
documentation for the project.

As a documentation generation tool we generally require a certain
consistency between generations of different version of the same
codebase, unless something in the code changed so that it justifies the
changes in the documentation itself.

This stability is a further requirement for QDoc's regression testing
suite, which is enforced in CI.

Due to changes between LLVM 15 and LLVM 16, the provider of Clang, the
acquisition of the stringified version of certain documented elements
are not stable between the two version, so that the current codebase
cannot provide that required stability.

To address the main cause of this issue,
`clang::TypeName::getFullyQualifiedName`, a function that provides a
stringified representation of a type's name that is fully qualified as
if the type was to be referenced from the end of the inspected
translation unit, can be used to retrieve certain elements so as to
avoid changes in QDoc output on LLVM 15 and LLVM 16.

QDoc currently redistributes a custom version of the header that defines
`clang::TypeName::getFullyQualifiedName`. This is to allow certain
behavior changes that makes the function not suitable for QDoc required
usage as a documentation generator tool.

In particular, when `clang::TypeName::getFullyQualifiedType` /
`clang::TypeName::getFullyQualifiedName` encounter a non-dependent
type/type alias that is defined under a templated class/struct, it
qualifies the type/type alias under a specialization of the templated
class, if any was declared.

That is, in:

```
template<typename T>
class Foo {
    using Bar = T;
}

using Baz = Foo<int>;
```

Usages of `Foo::Bar` will be qualified as `Foo<int>::Bar`.

When the using `Baz = Foo<int>;` line is removed, as there are would be
no specialization encountered in the translation unit, usages of
`Foo::Bar` would instead be qualified as `Foo<T>::Bar`.

When multiple specializations are present, the one that is chosen is the
first one that is encountered; due to the current implementation of the
behavior and the fact that iterating the specialization of a
`ClassTemplateDecl` respects the order in which the specializations were
stored in.

That is, if we were to add a reference to some other specialization in
the above example, so that it would be parsed before the current `using
Baz = Foo<int>;`, say:

```
template<typename T>
class Foo {
    using Bar = T;
}

using Bat = Foo<double>;
using Baz = Foo<int>;
```

Then usages of `Foo::Bar` would be qualified as `Foo<double>::Bar`
instead of `Foo<int>::Bar`.

Should the same declaration be added, instead, after the using `Baz =
Foo<int>;` line, then the qualification would remain consistent with our
previous version.

This behavior has a few issues for the usage that QDoc has for
`clang::TypeName::getFullyQualifiedName`.

Having the produced output change because of the first encountered
specialization changing in the codebase is not feasible, as it break
this desirable property of a documentation set and may break our
regression-testing suite that is enforced in CI.

Furthermore, the chosen specialization may produce incorrect
information, which we cannot show to the end user of the documentation
set.

For example, Qt's documentation currently has `QFuture` define a
`const_iterator` alias that is later used as the argument or return type
of some of our documented callables.

Due to the way Qt's codebase is set up, if QDoc were to use
`clang::TypeName::getFullyQualifiedName`, the chosen specialization
would be `QFuture<void>` so that, for example, `QFuture::constBegin()`
would be shown to the user as returning a
`QFuture<void>::const_iterator`. Nonetheless, `QFuture<void>` is the
only specialization that cannot have a `const_iterator` and, as a
consequence, doesn't have a `constBegin` method in the first place.

Even in the case where we could decide which specialization is chosen,
so as to avoid an outright erroneous documentation, the user of a
documentation set would still be generally confused by the provided
information.

For example, they may be convinced that `QList<T>::at` always returns a
`QList<int>::const_reference`.

To avoid those issues and provide a more consistent behavior that avoids
changing the output when relatively unrelated declarations are present
in the translation unit; `clang::TypeName::getFullyQualifiedType` /
`clang::TypeName::getFullyQualifiedName` will now qualify such types
using the original templated declaration instead of one of the
specializations.

Task-number: QTBUG-111580
Change-Id: I254065a7a32b8cc4d55017ff7467a2c4c04d6cd0
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>