QDoc: Fully qualify non-type template parameters
QDoc uses Clang's APIs, provided by the LLVM project, to parse the
source code of a project, both to extract the user-provided
documentation and to perform some sanity checks.

When parsing the source code, a user-presentable version of certain
documented elements that will appear in the output documentation, is
obtained through the use of Clang's API.

For example, when QDoc needs to show a type to the user, say because it
is showing the signature of a function, the representation of that type
that is shown to the user is acquired through the user of certain
Clang's APIs calls.

Due to certain changes between LLVM 15 and LLVM 16, the output of the
API calls that QDoc performs to obtain those representations is not
stable between versions.

QDoc has generally been leaning on the default behavior of certain
functions such as `clang_getTypeSpelling`.
The default behavior has been generally producing an output that was
acceptable for QDoc purposes and specifically to be shown to an end user
of a documentation set.

With the changes in LLVM 16, this default behavior changed so that the
output will generally not be particularly qualified (based on the
inspected codebase) nor will it be, necessarily, consistent (based on
the inspected codebase).

Furthermore, the changes generally break QDoc's regression testing suite
that is enforced in CI, making it so that it is not feasible to move
between versions of LLVM.

To avoid the issue, a more stable approach is required. In particular,
it is possible to use `clang::TypeName::getFullyQualfiedName`, of which
QDoc exports a custom version, to ensure that the representation of type
names extracted from a codebase is both stable and consistent.

When QDoc parses the source code, the template declarations for a
template class or function is retrieved.
This is later shown to the user in multiple places in the output
documentation.

For example, a template class `template<typename T> class Foo`, if
documented, would show the `template<typename T>` line as part of the
class documentation.

When QDoc retrieves the template declaration, if a non-type template
parameter is present, it will retrieve it by usages of
`clang_getTypeSpelling`, which incurs in the issue described above.

To avoid the issue, and ensure that non-type template parameters in a
template declaration are shown in a consistent and stable way, the
retrieval of non-type template parameters in QDoc was modified to use
`clang::TypeName::getFullyQualifiedName`.

To do so, `get_cursor_declaration`, the entry point to switch from Clang's C API
to Clang's C++ API, was moved at the top of "clangcodeparser.cpp", where
it previously resided, to ensure that it is in scope for usages in
various part of the file.

A new free function, `get_fully_qualified_type_name`, was added to
"clangcodeparser.cpp", taking care of wrapping
`clang::TypeName::getFullyQualifiedType` in a smaller interface and
providing a common entry point to the retrieval of type "spellings" for
QDoc.

The relevant section of `getTemplateParameters`, which executes the
retrieval of a presentable template declaration for a templated class or
function, was modified to avoid the usage of `clang_getTypeSpelling` for
non-type template parameters.

Instead, a non-type template parameter will now be lowered to a Clang's
C++ AST node and its type "spelling" will be retrieved by the usage of
`get_fully_qualified_type_name`.

This patch is generally be expected to have an effect on the output
documentation of certain projects, as non-type template parameters will
be consistently qualified while before they may of may not have been
qualified based on usage.

Task-number: QTBUG-111580
Change-Id: Iea6860a9ae90f79fb66bc301119e806dd66ccbe5
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>