QDoc: Fully qualify conversion operators
QDoc uses Clang's APIs, provided by the LLVM project, to parse the
source code of a project, both to extract the user-provided
documentation and to perform some sanity checks.

When parsing the source code, a user-presentable version of certain
documented elements that will appear in the output documentation, is
obtained through the use of Clang's API.

For example, when QDoc needs to show a type to the user, say because it
is showing the signature of a function, the representation of that type
that is shown to the user is acquired through the user of certain
Clang's APIs calls.

Due to certain changes between LLVM 15 and LLVM 16, the output of the
API calls that QDoc performs to obtain those representations is not
stable between versions.

QDoc has generally been leaning on the default behavior of certain
functions such as `clang_getTypeSpelling`.
The default behavior has been generally producing an output that was
acceptable for QDoc purposes and specifically to be shown to an end user
of a documentation set.

With the changes in LLVM 16, this default behavior changed so that the
output will generally not be particularly qualified (based on the
inspected codebase) nor will it be, necessarily, consistent (based on
the inspected codebase).

Furthermore, the changes generally break QDoc's regression testing suite
that is enforced in CI, making it so that it is not feasible to move
between versions of LLVM.

To avoid the issue, a more stable approach is required. In particular,
it is possible to use `clang::TypeName::getFullyQualfiedName`, of which
QDoc exports a custom version, to ensure that the representation of type
names extracted from a codebase is both stable and consistent.

When QDoc parses the source code and registers a conversion operator, it
extracts the type that the operator converts to so that it can later
show the operator in multiple places in the output documentation.

For example, some conversion operator to `int` in a class `Foo`, if
documented, would be shown as part of the member of the class `Foo`.

When QDoc retrieves the name of the conversion operator, it will
retrieve its "conversion type", to use as part of its name, with
`clang_getTypeSpelling`, which incurs in the issue described above.

To avoid the issue, and ensure types in conversion operators are shown
in a more consistent and stable way, the retrieval of the conversion
type for a conversion operator in QDoc was modified to use
`clang::TypeName::getFullyQualifiedName`.

The relevant section of `functionName`, which takes care of obtaining
the "name" of a callable that is later shown to the user, was modified
to avoid the usage of `clang_getTypeSpelling` for conversion operators.

Instead, a conversion operator will now be lowered to a Clang's C++ AST
node and its type "spelling" will be retrieved by the usage of
`get_fully_qualified_type_name`.

This patch is generally expected to have an effect on the output
documentation of certain projects, as the type encoded in the
declaration of a conversion operator will be consistently qualified
while before it may or may not have been qualified based on usage.

Due to a current limitation of the QDoc linking mechanism, this may
induce missing or reduced linking for certain projects, in particular
due to the possibility of template parameters appearing as part of a
conversion operator declaration.
This linking issue is expected to be resolved separately from this
patch.

Task-number: QTBUG-111580
Change-Id: I5072dfe39655d86e01664bb84759757996b5e58c
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>