QDoc: Fully qualify the return type of callables
QDoc uses Clang's APIs, provided by the LLVM project, to parse the
source code of a project, both to extract the user-provided
documentation and to perform some sanity checks.

When parsing the source code, a user-presentable version of certain
documented elements that will appear in the output documentation, is
obtained through the use of Clang's API.

For example, when QDoc needs to show a type to the user, say because it
is showing the signature of a function, the representation of that type
that is shown to the user is acquired through the user of certain
Clang's APIs calls.

Due to certain changes between LLVM 15 and LLVM 16, the output of the
API calls that QDoc performs to obtain those representations is not
stable between versions.

QDoc has generally been leaning on the default behavior of certain
functions such as `clang_getTypeSpelling`.
The default behavior has been generally producing an output that was
acceptable for QDoc purposes and specifically to be shown to an end user
of a documentation set.

With the changes in LLVM 16, this default behavior changed so that the
output will generally not be particularly qualified (based on the
inspected codebase) nor will it be, necessarily, consistent (based on
the inspected codebase).

Furthermore, the changes generally break QDoc's regression testing suite
that is enforced in CI, making it so that it is not feasible to move
between versions of LLVM.

To avoid the issue, a more stable approach is required. In particular,
it is possible to use `clang::TypeName::getFullyQualfiedName`, of which
QDoc exports a custom version, to ensure that the representation of type
names extracted from a codebase is both stable and consistent.

When QDoc parses the source code and registers the information of a
callable, it extracts the return type of that callable so that it can be
later shown to the user in multiple places in the output documentation.

For example, the return type of some `void foo()` in a class `Foo`, if
documented, would be shown as part of the member of the class `Foo`, as
part of the summary of its member on the class page and in its own
section in the detailed documentation of the class page.

When QDoc retrieves analyzes a callable, it will retrieve its return
type by a usage of `clang_getTypeSpelling`, which incurs in the issue
described above.

To avoid the issue, and ensure that return types of callables are shown
in a more consistent and stable way, the retrieval of the return type
was modified to use `clang::TypeName::getFullyQualifiedName`.

The relevant section of `processFunction`, which takes care of obtaining
most information about a callable that will later be shown to the user,
was modified to avoid the usage of `clang_getTypeSpelling` for
return types.

Instead, the return type will be obtained by lowering the callable's
cursor to a Clang's C++ AST node and its type "spelling" will be
retrieved by the usage of `get_fully_qualified_type_name`.

The code in `processFunction` that takes care of retrieving the return
type is thus moved a bit down in the function body, so as to get access
to the already lowered cursor that we already use for other types of
retrieval.

This patch is generally expected to have an effect on the output
documentation of certain projects, as the type encoded in the return
type of a callable declaration will be consistently qualified while
before it may or may not have been qualified based on usage.

Due to a current limitation of the QDoc linking mechanism, this may
induce missing or reduced linking for certain projects, in particular
due to the possibility of template parameters appearing as part of a
return type.
This linking issue is expected to be resolved separately from this
patch.

Task-number: QTBUG-111580
Change-Id: Ide90a3754bea787bf4ea9217f25d444cfa892db7
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>