QDoc: Use clang::Decl::getAsFunction
When working with `clang::Decl`s that are known to represent callables
but it is not known whether they are are templated or not, to obtain the
relevant `clang::FunctionDecl`, the template and non-template
declaration have to be treated slightly differently.

`clang::Decl::getAsFunction` takes care of extracting a
`clang::FunctionDecl` from a declaration, taking into account the
difference between a templated and non-templated declaration.

QDoc was performing this `clang::Decl` -> `clang::FunctionDecl` cast in
`processFunction` in "clangcodeparser.cpp", manually taking care of the
differences between a templated and non-templated declaration, which
both may appear in that code-path.

The code that QDoc uses is equivalent to the one in
`clang::Decl::getAsFunction`, and was used as the existence of the
method was unknown at the time that code was written.

Hence, instead of performing the required checks manually,
`processFunction` now directly uses `clang::Decl::getAsFunction` to
obtain a `clang::FunctionDecl` when required.

Change-Id: I9929bc6ef57be38a86d633559b7a68477d8fac0a
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>