QDoc: Fully qualify the types of the arguments of callables
QDoc uses Clang's APIs, provided by the LLVM project, to parse the
source code of a project, both to extract the user-provided
documentation and to perform some sanity checks.

When parsing the source code, a user-presentable version of certain
documented elements that will appear in the output documentation, is
obtained through the use of Clang's API.

For example, when QDoc needs to show a type to the user, say because it
is showing the signature of a function, the representation of that type
that is shown to the user is acquired through the user of certain
Clang's APIs calls.

Due to certain changes between LLVM 15 and LLVM 16, the output of the
API calls that QDoc performs to obtain those representations is not
stable between versions.

QDoc has generally been leaning on the default behavior of certain
functions such as `clang_getTypeSpelling`.
The default behavior has been generally producing an output that was
acceptable for QDoc purposes and specifically to be shown to an end user
of a documentation set.

With the changes in LLVM 16, this default behavior changed so that the
output will generally not be particularly qualified (based on the
inspected codebase) nor will it be, necessarily, consistent (based on
the inspected codebase).

Furthermore, the changes generally break QDoc's regression testing suite
that is enforced in CI, making it so that it is not feasible to move
between versions of LLVM.

To avoid the issue, a more stable approach is required. In particular,
it is possible to use `clang::TypeName::getFullyQualfiedName`, of which
QDoc exports a custom version, to ensure that the representation of type
names extracted from a codebase is both stable and consistent.

When QDoc parses the source code and registers the information of a
callable, it extracts the type of each argument for that callable,
later shown to the user in multiple places in the output documentation.

For example, the argument of some `void foo(int)` in a class `Foo`, if
the callable is documented, would be shown as part of the members of the
class `Foo`, as part of the summary of its member on the class page and
in its own section in the detailed documentation of the class page.

When QDoc retrieves a callable, it will retrieve its argument by usages
of `clang::QualType::getAsString`, which uses the same code path as
`clang_getTypeSpelling`, so that it incurs in the issue described above.

To avoid the issue, and ensure that type of the arguments of a callable
are shown in a more consistent and stable way, the types was modified to
use `clang::TypeName::getFullyQualifiedName`.

The relevant section of `processFunction`, which takes care of obtaining
most information about a callable that will later be shown to the user,
was modified to avoid the usage of `getAsString` for the types of the
arguments of the callable.

Instead, the types will now be obtained by the usage of
`get_fully_qualified_type_name` which wraps
`clang::TypeName::getFullyQualifiedName`.

When QDoc parses the documentation and source code it builds an internal
representation that is used by later parts of the codebase.
During the parsing, this internal representation is often used to
"match" a `CXCursor` that is extracted from the codebase, to check
whether we already registered some data or, especially, to find to which
piece of data an `\fn` command refers to.

This "matching" process is performed in `findNodeForCursor` in
'clangcodeparser.cpp'.

As part of the matching process, QDoc takes into account the argument
types that are stored in certain candidates in the internal
representation and those that can be extracted from the inspected
`CXCursor`.

The way in which we retrieve the the types from the `CXCursor` while
matching has to generally match how we retrieved the types for the
internal representation, as we might otherwise fail to match a valid
candidate due to a different "spelling".

Hence, `findNodeForCursor` now uses `get_fully_qualified_type_name` when
retrieving the types of arguments from a `CXCursor`.

Some spacing was added in the function body to make certain units more
easily to parse/recognize for a reader.

This patch is generally expected to have an effect on the output
documentation of certain projects, as the types encoded for the
arguments of a callable callable declaration will be consistently
qualified while before it may or may not have been qualified based on
usage.

Due to a current limitation of the QDoc linking mechanism, this may
induce missing or reduced linking for certain projects, in particular
due to the possibility of template parameters appearing as part of a the
type of an argument.
This linking issue is expected to be resolved separately from this
patch.

Task-number: QTBUG-111580
Change-Id: Iff234869d4b9262ce623caaff7ad4eb4357f79ba
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>