QDoc: Remove unnecessary computation in `findNodeForCursor`
When QDoc parses the documentation and source code it builds an internal
representation that is used by later parts of the codebase.

During the parsing, this internal representation is often used to
"match" a `CXCursor` that is extracted from the codebase, to check
whether we already registered some data or, especially, to find to which
piece of data an `\fn` command refers to.

This "matching" process is performed in `findNodeForCursor` in
'clangcodeparser.cpp'.

As part of the matching process, QDoc takes into account the argument
types that are stored in certain candidates in the internal
representation and those that can be extracted from the inspected
`CXCursor`.

When those types do not match, QDoc tries to dequalify their
representation, based on the parent tree of the callable that is being
matched, and match the less qualified version of those types.

If, when all parents are exhausted, the types still do not match, then
the candidate is considered to not match the `CXCursor`.

The dequalification occurs due to some of those types being extracted by
user-provided `\fn` commands, which might spell the types differently
to how they are spelled in the code, from which the to-match types have
been extracted.

Due to recent changes in QDoc, all types are now obtained in "fully
qualified" form.
This form will be unique for each type, inside the same translation
unit, and will should not depend on the way the user might have written
the type, originally, as long as the types are the same.

Hence, it is not required anymore for `findNodeForCursor` to check
different qualifications of the same types when matching.

Thus, the code that performed this process is removed.

Change-Id: Ic7c752941d5b85ae4438a7cab325a4ae82e81984
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>