QAbstractFormBuilder: Remove some unnecessary DomPropertyHash conversions
This is the beginning of a series of patches to remove strings from
class QFormBuilderStrings.

In some QAbstractFormBuilder functions, the property lists are
converted to a DomPropertyHash. In some cases, this was done for just
one lookup. To avoid these conversions; add a helper to search for a
DomProperty by name to QFormBuilderExtra.

Add another helper QFormBuilderExtra::getLayoutMargins() for
retrieving layout margins from DOM properties.

With that in place, some string constants from QFormBuilderStrings can
be removed.

Pick-to: 6.6
Task-number: QTBUG-118473
Change-Id: Ieaed9cf14989a85f23b18186abe803fc8396897c
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>