QDoc: Remove unused `adjustTypeName`
When QDoc parses the source code of a project, it extracts and store
various information from it, saving the required one in an internal
intermediate format.

When QDoc was obtaining a user-presentable version of the return type of
a callable or a user-presentable version of the type of an argument of a
callable, it would perform certain presentation transformation.

In particular, for types that were defined under some record that
contained the inspected callable, it would remove the name of the
record.

For example, given a class `Foo` with `using Bar = ...` and a method of
`Foo`, say, `void baz(Bar)`, QDoc would ensure that the type of the
first argument of `baz` would be presented as `Bar` instead of
`Foo::Bar`.

Due to recent changes to the way QDoc extracts and presents type names
that were performed as part of a compatibility effort for different LLVM
versions, most types are now presented in a "fully qualified form", thus
removing the need for the above transformation.

Hence, `adjustTypeName` in "clangcodeparser.cpp", which performed the
presented transformation, was left unused and is now removed from the
code-base as dead code.

While the current presentation for QDoc has no need for
`adjustTypeName`, something similar to it might be re-included at some
point if the "full qualified form" ends up being considered less clear
to the end-user of a documentation set.

If that will be the case, this patch should not be reverted and the
current form of `adjustTypeName` should not be re-introduced.
Instead, a similar effort should be written at the generator phase,
which takes care of the presentation issue for the final output.

Change-Id: I2a3566b5c5c73b5bc7cc0d94f69dc4062bc3b319
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>