QDoc: Use clang::ASTContext::getPrintingPolicy
QDoc uses Clang's C and C++ APIs to parse the source code of a project
and extract information about it that are then used to build the final
documentation.

When QDoc parses the source code, it extracts a user-presentable name
for various types, such as the return type and argument types of
callables, which are later presented to the user.

Due to the implementation of the process that is used to extract this
those type information, QDoc has to provide a `clang::PrintingPolicy`,
which is used by the underlying APIs to decide how to present certain
elements.

Currently, QDoc will construct one such printing policy for each call,
in "clangcodeparser.cpp", to `get_fully_qualified_type_name`, which is
used to obtain a user-presentable version of a type name.

The printing policy will be constructed by using the
`clang::LangOptions` obtained from an AST context, as suggested by the
API to construct a sensible `clang::PrintingPolicy`.

Nonetheless, a `clang::ASTContext`, which is required to obtain type
information by `get_fully_qualified_type_name`, already constructs a
single such `clang::PrintingPolicy` when it is built.

Hence, instead of constructing a custom printing policy,
`get_fully_qualified_type_name` will now extract a reference to the one
stored in the `clang::ASTContext` for the inspected translation unit,
which is equivalent to the custom built one, so that it constructs a
single instance per translation unit instead of one per call.

Change-Id: Icfcd88ed9df5cb699fc0d1df578c04ae3abc5d1f
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>