CMake: Fix the XXX_lupdate targets for Visual Studio generators
When using the VS generator, the foo_lupdate targets were run on every
build, which is neither expected nor wanted. The reason for this is a
pecularity in MSBuild and as a reaction to this, special behavior of
CMake: add_dependencies(A B) will enable B in the default build of the
solution - even if B is not triggered by the ALL target. See upstream
CMake issue #16668 for details.

We now work around this problem by building foo_lupdate externally as
PRE_BUILD step of update_translations.

Pick-to: 6.5 6.6
Fixes: QTBUG-115166
Change-Id: I33110450de6289f99954890ff8785ecbe67e5a29
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>