QDoc: Add missing default argument initializer in QualTypeNames.h
QDoc recently adopted a file from upstream LLVM, that provides utilities
to obtain the fully qualified form of a type.

The file exports `getFullyQualifiedName` and `getFullyQualifiedType`,
which are supposed to have their third argument, `bool
WithGlobalNsPrefix`, have a default argument of "false".

This default was lost when adopting the file, due to a manual error.

Hence, to synchronize to the correct version of the code, the correct
default initializer for the relevant arguments was added to
"QualTypeNames.h".

Change-Id: I2fd47062df03bae19f62121a8f07515c42ab0d26
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>