Show licenses from LICENSES directory in attribution documentation
It shouldn't matter whether a license file is explicitly set via
"LicenseFile(s)", or retrieved from the LICENSES directory based on
the SPDX expression. So let's resolve the license files from SPDX
before trying to load the content of the files so that they are
printed by the qdoc backend later on.

Pick-to: 6.5 6.6
Change-Id: I9ab31fae6f51c18718804bc58c162fdb05e49bde
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>