qdoc: Ensure linking works to overloaded functions in the global scope
When using the \relates command to associate a global function to
a type, namespace, or a header, QDoc sets the 'next overload'
pointer of that function to nullptr, in order to detach the function
from other potential overloads which might be documented under
another related type or left undocumented or \internal.

However, if the global function has overloads, QDoc also removed
it from the overload list of the global namespace's function map.
This had an unintended consequence; it was possible that only
undocumented overloads were left under the global namespace in
the Node tree. Undocumented functions are then marked \internal,
making it impossible to use the \l command to link to a function
in the global scope even though there is a documented overload
that was associated with a type using \relates.

To fix, instead of removing the FunctionNode from the overload
list, move it to the end of the list. This way, we keep the
linked list of overloads intact relative to the new parent
node that 'adopts' the function, but also keep the node
accessible from the global namespace.

Fixes: QTBUG-119146
Pick-to: 6.6 6.5
Change-Id: I2d30a51cc009674448f19f53f50e1520559f0a28
Reviewed-by: Luca Di Sera <luca.disera@qt.io>