QDoc: Add a regression suite for templated callables
QDoc provides a regression test, "generatedoutput", that, while
incomplete, is used to keep track of certain changes that are touch the
end-to-end output of a QDoc run.

When QDoc processes a project, it extract certain information from the
underlying code-base of the proejct.
The extracted information are then used to perform certain sanity checks
and to build the final output documentation.

One such information is the template declaration of certain templated
documentable elements in a C++ code-base.

Currently, "generatedoutput" provides only a very small surface to track
changes to template declarations.

While QDoc extracts template declaration, they are not meaningful when
performing any sanity checks.

For example, when QDoc tries to "match" an "\fn" command to its
underlying documentable element, as extracted from the code-base,
template declarations are not taken into account.

This, in turn, means that two callables that are only distinguishable by
their template declaration will be indistinguishable to QDoc, preventing
the user from correctly documenting them as separate elements.

As this issue will be addressed shortly, QDoc will require changes that
can modify the output documentation with regards to template
declarations of callable elements.

To help in evaluating the effect of those changes, the surface of
"generatedoutput" is now increased with a new "test",
"templatedCallables".

The new test, while not exhaustive due to it being manually mantained,
provides some relatively realistic specimens for most base cases of
templated callables, so that changes in output can be tracked.

A new directory, "templatedCallables", was added under the "testdata"
for "generatedoutput".

"templatedCallables" defines a QDoc project, based on the general
configurations for "generatedoutput", that documents a simple code-base
exposing a variety of callable elements that are templated.

The code-base for the project lives in two new files,
"templated_callables.h", which contains the specimens definitions, and
"templated_callables.cpp", which contains a documentation block for each
specimen in the project.

"tst_generatedoutput.cpp", the main implementation file for the test,
was updated to take into consideration the new project, as is required
by the updating process for "generatedoutput".

The initial regression files for the new project were generated and are
now part of the project, under "expected_output".

The regression files show a series of errors with regards to the
presentation of template declarations, such as missing parameter pack
information or default argument initializers.
Those errors have been in QDoc for a long time, thus why they are
represented in our regression suite, and are expected to be fixed in
certain upcoming changes related to template declarations in QDoc.

Task-number: QTBUG-118080
Task-number: QTBUG-117881
Change-Id: I21bfb7eba56d34d7872670fe208c2049301c9145
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>